+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - `params_` on fitted models is not longer an OrderedDict, but a Pandas Series
  - `GammaGammaFitter` can accept a `weights` argument now.
  - `customer_lifelime_value` in `GammaGamma` now accepts a frequency argument.
+ - fixed a bug that was causing `ParetoNBDFitter` to generate data incorrectly.
 
 ### 0.10.1
  - performance improvements to `generate_data.py` for large datasets #195